requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic>=2.5.2,<3.0.0
email-validator==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
gitpython==3.1.40
httpx==0.25.2
openai==1.3.5
anthropic==0.7.1
google-generativeai==0.3.1
mistralai==0.0.11
groq==0.4.1
reportlab==4.0.7
pillow>=10.4.0
pyotp==2.9.0
qrcode==7.4.2
redis==5.0.1
stripe==7.0.0
werkzeug==3.0.1
